+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Updated System.Diagnostics.DiagnosticSource to version 5.0.0-rc.1.20428.3
+* Added `GetTagValue` extension method on `Activity` for retrieving tag values
+  efficiently
+  ([#1221](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1221))
 
 ## 0.5.0-beta.2